Implement Sign In/Sign Up feature
Moving main endpoints from __init__ to main.py
